重构status及job

+++ internal/order/internal/web/vo.go
@@ -62,7 +62,7 @@ type RetrieveOrderStatusReq struct {
 }
 
 type RetrieveOrderStatusResp struct {
-	OrderStatus int64 `json:"status"`
+	OrderStatus uint8 `json:"status"`
 }
 
 // CompleteOrderReq 完成订单
@@ -102,7 +102,7 @@ type Order struct {
 	PaymentSN          string      `json:"paymentSn"`
 	OriginalTotalPrice int64       `json:"originalPrice"`
 	RealTotalPrice     int64       `json:"realPrice"`
-	Status             int64       `json:"status"`
+	Status             uint8       `json:"status"`
 	Items              []OrderItem `json:"items"`
 	Payments           []Payment   `json:"payments"`
 	Ctime              int64       `json:"ctime"`
